This commit completes user story US 06.01.01 and adds a navigation bar.
The goal of user story US 06.01.01 was for the user to have a profile with a unique username. To achieve this, the user can use an email and password to create an account and sign in to the app. In addition to this, a navigation bar has been added to the app to enable an easy switch from one activity to another.